fix(ram): resolve SRAM overflow on STM32G431CBU6 (32KB)
RAM was 50944 B / 32752 B = 155.54% (overflow 18192 B).
Four targeted reductions bring it within budget (~32568 B):

  - bsp_usart.h: USART_TXBUFF_SIZE 2048→512 (-7680 B, 5×1536)
  - config.h: MAX_POLE_PAIRS 20→7 (-2600 B, matches DEFAULT_POLE_PAIRS)
  - calibration_context.h: error_array int→int16_t (-4000 B)
  - FreeRTOSConfig.h: heap 12288→8192 (-4096 B)

Also included in this commit:
  - tim.c: dynamic ADC trigger offset via HW_PWM_ADC_TRIGGER_OFFSET_TICKS
  - linker scripts: mark .init_array/.fini_array READONLY
  - motor_data.c: remove spurious .calib_ctx={0} initializer (-Wmissing-braces)
  - bsp_log.c: remove double-buffer, safe \r\n, filename truncation
  - main.h: #ifndef guard on MCPWM_DEADTIME_CLOCKS (redefinition warning)
  - calib_encoder.c: #undef ENC before local shadow (redefinition warning)

+++ Src/ALGO/motor/calibration_context.h
@@ -98,8 +98,8 @@ typedef struct {
   float phase_set;         // Current electrical angle
   int16_t sample_count;    // Sample counter
   float next_sample_time;  // Next sampling timestamp
-  int error_array_storage[SAMPLES_PER_POLE_PAIR * MAX_POLE_PAIRS];
-  int *error_array;        // Error array (preallocated storage)
+  int16_t error_array_storage[SAMPLES_PER_POLE_PAIR * MAX_POLE_PAIRS];
+  int16_t *error_array;    // Error array (preallocated storage)
   size_t error_array_size; // Array size
 } EncoderCalibContext;
